pyproject.toml
[project]
name = "xtralogistics"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "arize-phoenix[evals]>=11.37.0",
    "bs4>=0.0.2",
    "fastapi>=0.120.1",
    "fitz>=0.0.1.dev2",
    "frontend>=0.0.3",
    "openai>=2.6.1",
    "openinference-instrumentation-bedrock>=0.1.28",
    "openinference-instrumentation-openai>=0.1.34",
    "opentelemetry-api>=1.37.0",
    "opentelemetry-exporter-otlp>=1.37.0",
    "opentelemetry-sdk>=1.37.0",
    "pdf2image>=1.17.0",
    "phoenix>=0.9.1",
    "psycopg2>=2.9.11",
    "pydantic>=2.12.3",
    "pymupdf>=1.26.5",
    "python-dotenv>=1.2.1",
    "requests>=2.32.5",
    "streamlit>=1.50.0",
    "uvicorn>=0.38.0",
]
